+++ client_secret.json
@@ -1,0 +1,14 @@
+{
+  "installed": {
+    "client_id": "520061647340-1d5q0ofr3gnftabsoi1isb621a3ooesj.apps.googleusercontent.com",
+    "project_id": "studious-set-197716",
+    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+    "token_uri": "https://accounts.google.com/o/oauth2/token",
+    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+    "client_secret": "25hbC60_Sf6wM1GJwTtxricP",
+    "redirect_uris": [
+      "urn:ietf:wg:oauth:2.0:oob",
+      "http://localhost"
+    ]
+  }
+}
